pyproject.toml
[project]
name = "agent-extract"
version = "0.1.0"
description = "Universal Document Intelligence Platform - High-accuracy multi-format document extraction with AI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Core dependencies
    "ipykernel>=6.30.1",
    "jupyter>=1.1.1",
    
    # AI/LLM dependencies
    "langchain>=0.3.0",
    "langchain-ollama>=0.2.0",
    "langgraph>=0.2.0",
    
    # Document parsing
    "pymupdf>=1.26.5",
    "pdfplumber>=0.11.0",
    "python-docx>=1.1.0",
    "openpyxl>=3.1.0",
    "pandas>=2.2.0",
    "unstructured>=0.10.0",
    
    # OCR and image processing
    "paddleocr>=2.7.0",
    "paddlepaddle>=2.6.0",
    "pytesseract>=0.3.10",
    "pdf2image>=1.17.0",
    "pillow>=11.3.0",
    "opencv-python>=4.9.0",
    
    # Table extraction
    "camelot-py[cv]>=0.11.0",
    "tabulate>=0.9.0",
    
    # Data validation and schemas
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",
    
    # CLI and output
    "typer>=0.12.0",
    "rich>=13.7.0",
    "click>=8.1.0",
    
    # Utilities
    "python-magic-bin>=0.4.14; sys_platform == 'win32'",
    "python-magic>=0.4.27; sys_platform != 'win32'",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
]
api = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.9",
]
all = [
    "agent-extract[dev,api]",
]

[project.scripts]
agent-extract = "agent_extract.cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ['py312']

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src/agent_extract --cov-report=term-missing"
